create BasicInfo components
